CAT-1932 Reduce amount of progress messages (only integer progress values)

+++ catroid/src/org/catrobat/catroid/scratchconverter/Client.java
@@ -60,7 +60,7 @@ public interface Client {
 		void onJobScheduled(Job job);
 		void onConversionReady(Job job);
 		void onConversionStart(Job job);
-		void onJobProgress(Job job, double progress);
+		void onJobProgress(Job job, short progress);
 		void onJobOutput(Job job, String[] lines);
 		void onConversionFinished(Job job, DownloadFinishedCallback downloadFinishedCallback, String downloadURL,
 				Date cachedDate);
