+++ catroid/src/org/catrobat/catroid/ui/scratchconverter/JobViewListener.java
@@ -31,7 +31,7 @@ public interface JobViewListener {
 	void onJobScheduled(Job job);
 	void onJobReady(Job job);
 	void onJobStarted(Job job);
-	void onJobProgress(Job job, double progress);
+	void onJobProgress(Job job, short progress);
 	void onJobOutput(Job job, @NonNull String[] lines);
 	void onJobFinished(Job job);
 	void onJobFailed(Job job);
